pyproject.toml
[tool.poetry]
name = "drizm-django-commons"
version = "0.2.0"
description = "Django commons for the Drizm organization"
authors = ["ThaRising <kochbe.ber@gmail.com>"]
maintainers = ["Dominik Lewandowski <dominik.lewandow@gmail.com>"]

license = "MIT"
readme = "README.md"
homepage = "https://github.com/drizm-team/django-commons"
repository = "https://github.com/drizm-team/django-commons"
keywords = ["commons", "django"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content"
]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.1.1"
djangorestframework = "3.12.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
drizm-commons = "^0.2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
